src/main/webapp/app/entities/mise-en-observation/route/mise-en-observation-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IMiseEnObservation } from '../mise-en-observation.model';
import { MiseEnObservationService } from '../service/mise-en-observation.service';

import { MiseEnObservationRoutingResolveService } from './mise-en-observation-routing-resolve.service';

describe('MiseEnObservation routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: MiseEnObservationRoutingResolveService;
  let service: MiseEnObservationService;
  let resultMiseEnObservation: IMiseEnObservation | null | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(MiseEnObservationRoutingResolveService);
    service = TestBed.inject(MiseEnObservationService);
    resultMiseEnObservation = undefined;
  });

  describe('resolve', () => {
    it('should return IMiseEnObservation returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultMiseEnObservation = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultMiseEnObservation).toEqual({ id: 123 });
    });

    it('should return null if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultMiseEnObservation = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultMiseEnObservation).toEqual(null);
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse<IMiseEnObservation>({ body: null })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultMiseEnObservation = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultMiseEnObservation).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
